Move MusicMemoryActivity to screens.main.musicmemory.view
